requirements.txt
mcp[cli]
httpx
duckduckgo-search
langchain-community
langchain-groq
langchain-mcp-adapters
langgraph
pandas
redis[search]
sentence-transformers
python-dotenv
fastapi
uvicorn